Wrapper around EnhancedVolcano::EnhancedVolcano()

Wrapper around EnhancedVolcano::EnhancedVolcano()

Usage

EnhancedVolcano2(
  toptable,
  lab,
  x,
  y,
  pCutoff = 0.001,
  FCcutoff = 1.5,
  col_by_regul = TRUE,
  col_up = "#b2182b",
  col_down = "#2166ac",
  col_others = "#e0e0e0",
  pCutoffCol = NULL,
  ...
)

Arguments

toptable

toptable EnhancedVolcano::EnhancedVolcano()

lab

lab EnhancedVolcano::EnhancedVolcano()

x

x EnhancedVolcano::EnhancedVolcano()

y

y EnhancedVolcano::EnhancedVolcano()

pCutoff

pCutoff EnhancedVolcano::EnhancedVolcano()

FCcutoff

FCcutoff EnhancedVolcano::EnhancedVolcano()

col_by_regul

lgl, default TRUE, whether to color variable by gene regulation - Up, Down, Other

col_up

a character string denoting color for up genes, works only when col_by_regul = TRUE.

col_down

a character string denoting color for down genes, works only when col_by_regul = TRUE.

col_others

a character string denoting color for other genes, works only when col_by_regul = TRUE.

pCutoffCol

character string denoting the column of p-values on which pCutoff should be applied. Defaults to the column used for y.

...

Other parameters pass to EnhancedVolcano::EnhancedVolcano()

Value

a volcano plot