Filter selected orgdb columns by gene-type

Filter selected orgdb columns by gene-type

Usage

.filter_orgdb_by_genetype(
  orgdb,
  col_genetype = "GENETYPE",
  feature_type = "protein-coding",
  col_return = "ENTREZID"
)

Arguments

orgdb

an OrgDB

col_genetype

a column name to filter by the value given in feature_type

feature_type

a string denoting value to filter the column mentioned by col_genetype.

col_return

a character string denoting a column to return from an OrgDB.

Value

a character string of the column mentioned in col_return.

Examples

.filter_orgdb_by_genetype(orgdb = org.Hs.eg.db::org.Hs.eg.db ,col_return = "ENTREZID",feature_type = "snRNA")
#> 'select()' returned 1:1 mapping between keys and columns
#>   [1] "6060"      "6063"      "6066"      "26824"     "26827"     "26828"    
#>   [7] "26829"     "26830"     "26831"     "26832"     "26834"     "26835"    
#>  [13] "26863"     "26864"     "26869"     "26870"     "26871"     "125050"   
#>  [19] "267010"    "100126780" "100126781" "100126798" "100126799" "100126800"
#>  [25] "100147744" "100151683" "100151684" "100169951" "100169952" "100169953"
#>  [31] "100169954" "100169955" "100169956" "100169957" "100169958" "100169959"
#>  [37] "100170216" "100170217" "100170218" "100170219" "100170220" "100170221"
#>  [43] "100170222" "100170223" "100170224" "100170225" "100170226" "100170227"
#>  [49] "100170228" "100191063" "101447996" "101954264" "101954266" "101954267"
#>  [55] "101954268" "101954269" "101954271" "101954272" "101954273" "101954275"
#>  [61] "101954276" "101954277" "101954278" "103625684" "115409981" "115409982"
#>  [67] "115409983" "115409984" "115409988" "115409989" "115482718" "124900488"
#>  [73] "124900516" "124900632" "124900883" "124900884" "124900885" "124900887"
#>  [79] "124900889" "124900915" "124901190" "124901191" "124901192" "124901219"
#>  [85] "124901224" "124901516" "124901518" "124901522" "124901523" "124901863"
#>  [91] "124902076" "124902077" "124902331" "124902333" "124902335" "124902818"
#>  [97] "124902821" "124903091" "124903094" "124903096" "124903097" "124903120"
#> [103] "124903275" "124903414" "124903782" "124904107" "124904108" "124904109"
#> [109] "124904111" "124904112" "124904122" "124904135" "124904136" "124904137"
#> [115] "124904138" "124904139" "124904140" "124904141" "124904142" "124904143"
#> [121] "124904144" "124904145" "124904146" "124904147" "124904364" "124904613"
#> [127] "124904616" "124904619" "124904621" "124904624" "124904627" "124904631"
#> [133] "124904634" "124904651" "124904694" "124904697" "124904706" "124904727"
#> [139] "124904732" "124904741" "124904748" "124904756" "124904769" "124904986"
#> [145] "124905062" "124905168" "124905169" "124905170" "124905264" "124905265"
#> [151] "124905307" "124905313" "124905321" "124905326" "124905327" "124905334"
#> [157] "124905519" "124905522" "124905573" "124905574" "124905808" "124905809"
#> [163] "124906135" "124906137" "124906338" "124906342" "124906343" "124906683"
#> [169] "124907960" "124907961" "124907963" "124907964" "124907965" "124908124"
.filter_orgdb_by_genetype(orgdb = org.Hs.eg.db::org.Hs.eg.db ,col_return = "ENTREZID",feature_type = "protein-coding")
#> 'select()' returned 1:1 mapping between keys and columns
#>    [1] "1"    "2"    "9"    "10"   "12"   "13"   "14"   "15"   "16"   "18"  
#>   [11] "19"   "20"   "21"   "22"   "23"   "24"   "25"   "26"   "27"   "28"  
#>   [21] "29"   "30"   "31"   "32"   "33"   "34"   "35"   "36"   "37"   "38"  
#>   [31] "39"   "40"   "41"   "43"   "47"   "48"   "49"   "50"   "51"   "52"  
#>   [41] "53"   "54"   "55"   "56"   "58"   "59"   "60"   "70"   "71"   "72"  
#>   [51] "81"   "86"   "87"   "88"   "89"   "90"   "91"   "92"   "93"   "94"  
#>   [61] "95"   "97"   "98"   "100"  "101"  "102"  "103"  "104"  "105"  "107" 
#>   [71] "108"  "109"  "111"  "112"  "113"  "114"  "115"  "116"  "117"  "118" 
#>   [81] "119"  "120"  "123"  "124"  "125"  "126"  "127"  "128"  "130"  "131" 
#>   [91] "132"  "133"  "134"  "135"  "136"  "140"  "141"  "142"  "143"  "146" 
#>  [101] "147"  "148"  "150"  "151"  "152"  "153"  "154"  "155"  "156"  "157" 
#>  [111] "158"  "159"  "160"  "161"  "162"  "163"  "164"  "165"  "166"  "167" 
#>  [121] "173"  "174"  "175"  "176"  "177"  "178"  "181"  "182"  "183"  "185" 
#>  [131] "186"  "187"  "189"  "190"  "191"  "196"  "197"  "199"  "202"  "203" 
#>  [141] "204"  "205"  "207"  "208"  "210"  "211"  "212"  "213"  "214"  "215" 
#>  [151] "216"  "217"  "218"  "219"  "220"  "221"  "222"  "223"  "224"  "225" 
#>  [161] "226"  "229"  "230"  "231"  "238"  "239"  "240"  "241"  "242"  "246" 
#>  [171] "247"  "248"  "249"  "250"  "251"  "257"  "258"  "259"  "262"  "265" 
#>  [181] "266"  "267"  "268"  "269"  "270"  "271"  "272"  "273"  "274"  "275" 
#>  [191] "276"  "277"  "278"  "279"  "280"  "283"  "284"  "285"  "286"  "287" 
#>  [201] "288"  "290"  "291"  "292"  "293"  "301"  "302"  "306"  "307"  "308" 
#>  [211] "309"  "310"  "311"  "312"  "313"  "314"  "316"  "317"  "318"  "319" 
#>  [221] "320"  "321"  "322"  "323"  "324"  "325"  "326"  "327"  "328"  "329" 
#>  [231] "330"  "331"  "332"  "333"  "334"  "335"  "336"  "337"  "338"  "339" 
#>  [241] "341"  "343"  "344"  "345"  "346"  "347"  "348"  "350"  "351"  "353" 
#>  [251] "354"  "355"  "356"  "357"  "358"  "359"  "360"  "361"  "362"  "363" 
#>  [261] "364"  "366"  "367"  "368"  "369"  "372"  "373"  "374"  "375"  "377" 
#>  [271] "378"  "379"  "381"  "382"  "383"  "384"  "387"  "388"  "389"  "390" 
#>  [281] "391"  "392"  "393"  "394"  "395"  "396"  "397"  "398"  "399"  "400" 
#>  [291] "401"  "402"  "403"  "405"  "406"  "407"  "408"  "409"  "410"  "411" 
#>  [301] "412"  "414"  "415"  "416"  "417"  "419"  "420"  "421"  "427"  "429" 
#>  [311] "430"  "432"  "433"  "434"  "435"  "438"  "439"  "440"  "443"  "444" 
#>  [321] "445"  "460"  "462"  "463"  "466"  "467"  "468"  "471"  "472"  "473" 
#>  [331] "474"  "475"  "476"  "477"  "478"  "479"  "480"  "481"  "482"  "483" 
#>  [341] "486"  "487"  "488"  "489"  "490"  "491"  "492"  "493"  "495"  "496" 
#>  [351] "498"  "501"  "506"  "509"  "513"  "514"  "515"  "516"  "517"  "518" 
#>  [361] "521"  "522"  "523"  "525"  "526"  "527"  "528"  "529"  "533"  "534" 
#>  [371] "535"  "537"  "538"  "539"  "540"  "545"  "546"  "547"  "549"  "550" 
#>  [381] "551"  "552"  "553"  "554"  "558"  "563"  "566"  "567"  "570"  "571" 
#>  [391] "572"  "573"  "574"  "575"  "576"  "577"  "578"  "579"  "580"  "581" 
#>  [401] "582"  "583"  "585"  "586"  "587"  "590"  "593"  "594"  "595"  "596" 
#>  [411] "597"  "598"  "599"  "602"  "604"  "605"  "607"  "608"  "610"  "611" 
#>  [421] "613"  "617"  "622"  "623"  "624"  "627"  "629"  "631"  "632"  "633" 
#>  [431] "634"  "635"  "636"  "637"  "638"  "639"  "640"  "641"  "642"  "643" 
#>  [441] "644"  "645"  "646"  "648"  "649"  "650"  "651"  "652"  "653"  "654" 
#>  [451] "655"  "656"  "657"  "658"  "659"  "660"  "661"  "662"  "663"  "664" 
#>  [461] "665"  "666"  "667"  "668"  "669"  "670"  "671"  "672"  "673"  "675" 
#>  [471] "676"  "677"  "678"  "680"  "682"  "683"  "684"  "685"  "686"  "687" 
#>  [481] "688"  "689"  "694"  "695"  "696"  "699"  "701"  "705"  "706"  "708" 
#>  [491] "710"  "712"  "713"  "714"  "715"  "716"  "717"  "718"  "719"  "720" 
#>  [501] "721"  "722"  "725"  "726"  "727"  "728"  "729"  "730"  "731"  "732" 
#>  [511] "733"  "734"  "735"  "738"  "740"  "741"  "744"  "745"  "746"  "747" 
#>  [521] "752"  "753"  "754"  "755"  "757"  "758"  "759"  "760"  "761"  "762" 
#>  [531] "763"  "765"  "766"  "767"  "768"  "770"  "771"  "773"  "774"  "775" 
#>  [541] "776"  "777"  "778"  "779"  "780"  "781"  "782"  "783"  "784"  "785" 
#>  [551] "786"  "788"  "790"  "793"  "794"  "795"  "796"  "797"  "799"  "800" 
#>  [561] "801"  "805"  "808"  "810"  "811"  "813"  "814"  "815"  "816"  "817" 
#>  [571] "818"  "819"  "820"  "821"  "822"  "823"  "824"  "825"  "826"  "827" 
#>  [581] "828"  "829"  "830"  "831"  "832"  "833"  "834"  "835"  "836"  "837" 
#>  [591] "838"  "839"  "840"  "841"  "842"  "843"  "844"  "845"  "846"  "847" 
#>  [601] "857"  "858"  "859"  "860"  "861"  "862"  "863"  "864"  "865"  "866" 
#>  [611] "867"  "868"  "869"  "871"  "873"  "874"  "875"  "881"  "883"  "885" 
#>  [621] "886"  "887"  "889"  "890"  "891"  "892"  "894"  "896"  "898"  "899" 
#>  [631] "900"  "901"  "902"  "904"  "905"  "908"  "909"  "910"  "911"  "912" 
#>  [641] "913"  "914"  "915"  "916"  "917"  "919"  "920"  "921"  "922"  "923" 
#>  [651] "924"  "925"  "926"  "927"  "928"  "929"  "930"  "931"  "932"  "933" 
#>  [661] "939"  "940"  "941"  "942"  "943"  "944"  "945"  "946"  "947"  "948" 
#>  [671] "949"  "950"  "951"  "952"  "953"  "954"  "955"  "956"  "957"  "958" 
#>  [681] "959"  "960"  "961"  "962"  "963"  "965"  "966"  "967"  "968"  "969" 
#>  [691] "970"  "971"  "972"  "973"  "974"  "975"  "976"  "977"  "978"  "983" 
#>  [701] "984"  "987"  "988"  "989"  "990"  "991"  "993"  "994"  "995"  "996" 
#>  [711] "997"  "998"  "999"  "1000" "1001" "1002" "1003" "1004" "1005" "1006"
#>  [721] "1007" "1008" "1009" "1010" "1012" "1013" "1014" "1015" "1016" "1017"
#>  [731] "1018" "1019" "1020" "1021" "1022" "1024" "1025" "1026" "1027" "1028"
#>  [741] "1029" "1030" "1031" "1032" "1033" "1036" "1039" "1040" "1041" "1043"
#>  [751] "1044" "1045" "1046" "1047" "1048" "1050" "1051" "1052" "1053" "1054"
#>  [761] "1056" "1058" "1059" "1060" "1062" "1063" "1066" "1068" "1069" "1070"
#>  [771] "1071" "1072" "1073" "1075" "1080" "1081" "1082" "1084" "1087" "1088"
#>  [781] "1089" "1101" "1102" "1103" "1104" "1105" "1106" "1107" "1108" "1109"
#>  [791] "1111" "1112" "1113" "1114" "1116" "1117" "1118" "1119" "1120" "1121"
#>  [801] "1122" "1123" "1124" "1128" "1129" "1130" "1131" "1132" "1133" "1134"
#>  [811] "1135" "1136" "1137" "1138" "1139" "1140" "1141" "1142" "1143" "1144"
#>  [821] "1145" "1146" "1147" "1149" "1152" "1153" "1154" "1155" "1158" "1159"
#>  [831] "1160" "1161" "1163" "1164" "1173" "1174" "1175" "1176" "1178" "1179"
#>  [841] "1180" "1181" "1182" "1183" "1184" "1185" "1186" "1187" "1188" "1191"
#>  [851] "1192" "1193" "1195" "1196" "1198" "1200" "1201" "1203" "1207" "1208"
#>  [861] "1209" "1211" "1212" "1213" "1215" "1230" "1232" "1233" "1234" "1235"
#>  [871] "1236" "1237" "1238" "1240" "1241" "1244" "1258" "1259" "1260" "1261"
#>  [881] "1262" "1263" "1264" "1265" "1266" "1267" "1268" "1269" "1270" "1271"
#>  [891] "1272" "1277" "1278" "1280" "1281" "1282" "1284" "1285" "1286" "1287"
#>  [901] "1288" "1289" "1290" "1291" "1292" "1293" "1294" "1295" "1296" "1297"
#>  [911] "1298" "1299" "1300" "1301" "1302" "1303" "1305" "1306" "1307" "1308"
#>  [921] "1310" "1311" "1312" "1314" "1315" "1316" "1317" "1318" "1325" "1326"
#>  [931] "1327" "1329" "1337" "1339" "1340" "1345" "1346" "1347" "1349" "1350"
#>  [941] "1351" "1352" "1353" "1355" "1356" "1357" "1358" "1359" "1360" "1361"
#>  [951] "1362" "1363" "1364" "1365" "1366" "1368" "1369" "1370" "1371" "1373"
#>  [961] "1374" "1375" "1376" "1378" "1379" "1380" "1381" "1382" "1384" "1385"
#>  [971] "1386" "1387" "1388" "1389" "1390" "1392" "1393" "1394" "1395" "1396"
#>  [981] "1397" "1398" "1399" "1400" "1401" "1404" "1406" "1407" "1408" "1409"
#>  [991] "1410" "1411" "1412" "1413" "1414" "1415" "1417" "1418" "1419" "1420"
#>  [ reached getOption("max.print") -- omitted 19598 entries ]
.filter_orgdb_by_genetype(orgdb = org.Hs.eg.db::org.Hs.eg.db ,col_return = "ENTREZID",feature_type = "awdfd")
#> 'select()' returned 1:1 mapping between keys and columns
#> character(0)