Get an input matrix for diff ASE heatmap

Get an input matrix for diff ASE heatmap

Usage

.get_aes_heatmap_data(
  se,
  event_names,
  samples,
  replicates = TRUE,
  method = "PSI",
  column_condition = "condition"
)

Arguments

se

an object of class NxtSE.

event_names

a character vector denoting valid ASE names to plot in heatmap.

samples

a character vector denoting valid sample names to plot in the heatmap.

replicates

logical, default TRUE, denoting whether to show sample replicates or not.

method

a character string denoting a method for data value. Can be one of the "PSI", "logit" or "Z-Score"

column_condtion

a name of column in se storing condition. Deafault: "condition"

Value

a data matrix

Examples

se <- SpliceWiz::SpliceWiz_example_NxtSE(novelSplicing = TRUE)
SpliceWiz::colData(se)$treatment <- rep(c("A", "B"), each = 3)
SpliceWiz::colData(se)$replicate <- rep(c("P","Q","R"), 2)
res_edgeR <- SpliceWiz::ASE_edgeR(se, "treatment", "A", "B", useQL = FALSE)
#> Feb 27 15:01:18 Performing edgeR contrast for included / excluded counts separately
#> Feb 27 15:01:19 Performing edgeR contrast for included / excluded counts together
event_names = get_edger_diff_ase_event_names(res_edgeR,   padj_cutoff = 1)
.get_aes_heatmap_data(se, event_names , samples = c("A", "B"), column_condition ="treatment")
#>                                               02H003       02H025      02H026
#> SE:TRA2B-205-exon2;TRA2B-203-int1        0.939990162 0.9234808702 0.963950432
#> SE:TRA2B-novelTr014-exon2;TRA2B-214-int2 0.002601908 0.0000000000 0.005889282
#> SE:SRSF3-203-exon5;SRSF3-novelTr003-int1 0.998073218 0.9969712267 0.997366342
#> SE:TRA2A-201-exon6;TRA2A-novelTr001-int1 0.983246073 0.9514348786 0.866251944
#> SE:TRA2B-205-exon4;TRA2B-novelTr003-int1 0.976542580 0.9924050633 0.984965731
#> SE:NSUN5-206-exon5;NSUN5-novelTr001-int1 0.968503937 0.9804878049 0.974025974
#> SE:NSUN5-201-exon6;NSUN5-novelTr002-int1 0.988165680 0.9801324503 0.995876289
#> SE:TRA2B-208-exon3;TRA2B-novelTr003-int1 0.959001783 0.9874037789 0.974455297
#> SE:SRSF1-207-exon3;SRSF1-novelTr001-int1 0.508867667 0.5181058496 0.899916597
#> SE:SRSF1-208-exon4;SRSF1-novelTr001-int1 0.517426273 0.5417218543 0.900497512
#> SE:TP53-206-exon10;TP53-201-int9         0.114503817 0.0375939850 0.023771791
#> SE:TP53-201-exon10;TP53-217-int1         1.000000000 1.0000000000 1.000000000
#> SE:TP53-204-exon10;TP53-201-int9         0.183098592 0.0724637681 0.034482759
#> SE:TRA2A-201-exon4;TRA2A-novelTr002-int1 0.977459016 0.9835306227 0.978260870
#> SE:SRSF2-novelTr004-exon2;SRSF2-209-int1 0.082018927 0.0339256866 0.179810726
#> SE:SRSF3-204-exon3;SRSF3-202-int2        0.000304971 0.0002959455 0.000000000
#> SE:TRA2A-212-exon2;TRA2A-201-int1        0.043010753 0.0325581395 0.038525963
#> SE:SRSF3-202-exon5;SRSF3-novelTr004-int1 0.999507874 1.0000000000 0.999749781
#> SE:SRSF1-201-exon2;SRSF1-novelTr004-int1 1.000000000 0.9941520468 0.998353909
#> SE:TP53-201-exon4;TP53-214-int3          1.000000000 1.0000000000 1.000000000
#> SE:TP53-207-exon2;TP53-201-int1          0.000000000 0.0000000000 0.000000000
#>                                              02H033       02H043       02H046
#> SE:TRA2B-205-exon2;TRA2B-203-int1        0.98191533 0.9796484736 0.9835514019
#> SE:TRA2B-novelTr014-exon2;TRA2B-214-int2 0.05340700 0.0018214936 0.0027598896
#> SE:SRSF3-203-exon5;SRSF3-novelTr003-int1 1.00000000 1.0000000000 0.9985465116
#> SE:TRA2A-201-exon6;TRA2A-novelTr001-int1 0.98347107 0.9813519814 0.9708994709
#> SE:TRA2B-205-exon4;TRA2B-novelTr003-int1 0.99384931 1.0000000000 0.9887926643
#> SE:NSUN5-206-exon5;NSUN5-novelTr001-int1 0.99650960 0.9925373134 0.9851851852
#> SE:NSUN5-201-exon6;NSUN5-novelTr002-int1 0.99778024 0.9951100244 1.0000000000
#> SE:TRA2B-208-exon3;TRA2B-novelTr003-int1 0.98927135 1.0000000000 0.9796860572
#> SE:SRSF1-207-exon3;SRSF1-novelTr001-int1 0.89321129 0.6630669546 0.8076502732
#> SE:SRSF1-208-exon4;SRSF1-novelTr001-int1 0.89377845 0.6822810591 0.8103448276
#> SE:TP53-206-exon10;TP53-201-int9         0.00754717 0.0190735695 0.0328358209
#> SE:TP53-201-exon10;TP53-217-int1         1.00000000 0.9898477157 1.0000000000
#> SE:TP53-204-exon10;TP53-201-int9         0.02139535 0.0270270270 0.0540145985
#> SE:TRA2A-201-exon4;TRA2A-novelTr002-int1 0.98673300 0.9891774892 0.9874213836
#> SE:SRSF2-novelTr004-exon2;SRSF2-209-int1 0.15789474 0.0085679315 0.0696132597
#> SE:SRSF3-204-exon3;SRSF3-202-int2        0.00000000 0.0007265682 0.0005821851
#> SE:TRA2A-212-exon2;TRA2A-201-int1        0.07373272 0.0417827298 0.0550335570
#> SE:SRSF3-202-exon5;SRSF3-novelTr004-int1 0.99942079 1.0000000000 0.9992411307
#> SE:SRSF1-201-exon2;SRSF1-novelTr004-int1 1.00000000 1.0000000000 0.9866412214
#> SE:TP53-201-exon4;TP53-214-int3          1.00000000 1.0000000000 1.0000000000
#> SE:TP53-207-exon2;TP53-201-int1          0.00000000 0.0000000000 0.0000000000