Get an input matrix for diff ASE heatmap

Get an input matrix for diff ASE heatmap

Usage

.get_ase_heatmap_data(
  se,
  event_names,
  samples,
  replicates = TRUE,
  method = "PSI",
  column_condition = "condition"
)

Arguments

se

an object of class NxtSE.

event_names

a character vector denoting valid ASE names to plot in the heatmap.

samples

a character vector denoting valid sample names to plot in the heatmap.

replicates

logical, default TRUE, denoting whether to show sample replicates or not.

method

a character string denoting a method for data value. Can be one of the "PSI", "logit" or "Z-Score"

column_condtion

a name of column in se storing condition. Deafault: "condition"

Value

a data matrix

Examples

se <- SpliceWiz::SpliceWiz_example_NxtSE(novelSplicing = TRUE)
SpliceWiz::colData(se)$treatment <- rep(c("A", "B"), each = 3)
SpliceWiz::colData(se)$replicate <- rep(c("P","Q","R"), 2)
res <- run_ase_diff_analysis(x = se, test_factor = "treatment", test_nom = "A" ,test_denom = "B",  IRmode ="annotated",  cutoff_lfc = 0.6, cutoff_padj = 1, regul_based_upon = 2)
#> Mar 22 17:19:27 Performing edgeR contrast for included / excluded counts separately
#> Mar 22 17:19:29 Performing edgeR contrast for included / excluded counts together
event_names = get_ASEsets_by_regulation(x = res, sample_comparisons = "A_VS_B", regul = "all") %>% unlist()
.get_ase_heatmap_data(se, event_names , samples = c("A", "B"), column_condition ="treatment")
#>                                                          02H003       02H025
#> AFE:TRA2B-215-exon1;TRA2B-203-exon1                0.9632308620 0.9504854369
#> SE:TRA2B-205-exon2;TRA2B-203-int1                  0.9399901623 0.9234808702
#> SE:TRA2B-novelTr014-exon2;TRA2B-214-int2           0.0026019081 0.0000000000
#> SE:SRSF3-203-exon5;SRSF3-novelTr003-int1           0.9980732177 0.9969712267
#> A5SS:SRSF2-202-exon1;SRSF2-205-exon1               0.9941348974 0.9970674487
#> SE:TRA2B-205-exon4;TRA2B-novelTr003-int1           0.9765425803 0.9924050633
#> A3SS:SRSF2-novelTr003-exon2;SRSF2-202-exon3        0.0000000000 0.0027223230
#> SE:TRA2A-201-exon6;TRA2A-novelTr001-int1           0.9832460733 0.9514348786
#> SE:NSUN5-206-exon5;NSUN5-novelTr001-int1           0.9685039370 0.9804878049
#> SE:NSUN5-201-exon6;NSUN5-novelTr002-int1           0.9881656805 0.9801324503
#> SRSF1/ENST00000584773_Intron4/known-exon           0.6867522681 0.6979458973
#> SE:TRA2B-208-exon3;TRA2B-novelTr003-int1           0.9590017825 0.9874037789
#> A3SS:SRSF1-201-exon4;SRSF1-novelTr001-exon2        0.6697247706 0.6802218115
#> SE:SRSF1-207-exon3;SRSF1-novelTr001-int1           0.5088676671 0.5181058496
#> A5SS:TRA2A-novelTr003-exon1;TRA2A-202-exon3        0.0000000000 0.0416666667
#> SE:SRSF1-208-exon4;SRSF1-novelTr001-int1           0.5174262735 0.5417218543
#> A5SS:SRSF1-206-exon1;SRSF1-201-exon1               0.0028571429 0.0027548209
#> A5SS:SRSF2-211-exon1;SRSF2-202-exon2               0.0108401084 0.0090171326
#> A5SS:TRA2A-208-exon2;TRA2A-202-exon3               0.6862745098 0.6034482759
#> TRA2B/ENST00000453386_Intron4/clean                0.0237505328 0.0121386425
#> A5SS:SRSF2-211-exon1;SRSF2-206-exon2               0.0289330922 0.0182481752
#> SE:SRSF3-204-exon3;SRSF3-202-int2                  0.0003049710 0.0002959455
#> SE:TRA2A-201-exon4;TRA2A-novelTr002-int1           0.9774590164 0.9835306227
#> SE:TRA2A-212-exon2;TRA2A-201-int1                  0.0430107527 0.0325581395
#> AFE:TRA2B-214-exon1;TRA2B-202-exon1                0.0000000000 0.0079365079
#> A3SS:TRA2A-201-exon6;TRA2A-novelTr006-exon2        0.9375000000 0.9500000000
#> SE:TP53-201-exon4;TP53-214-int3                    1.0000000000 1.0000000000
#> A3SS:TP53-201-exon5;TP53-218-exon5                 1.0000000000 1.0000000000
#> A3SS:TP53-201-exon2;TP53-205-exon2                 1.0000000000 1.0000000000
#> NSUN5/ENST00000252594_Intron2/clean                0.2072388128 0.2203024654
#> A3SS:TRA2B-208-exon3;TRA2B-205-exon4               0.0017814727 0.0021164021
#> A5SS:TRA2B-214-exon2;TRA2B-novelTr008-exon1        1.0000000000 1.0000000000
#> A3SS:NSUN5-201-exon3;NSUN5-203-exon3               1.0000000000 1.0000000000
#> A5SS:SRSF3-202-exon5;SRSF3-novelTr007-exon1        1.0000000000 0.9995389580
#> A3SS:SRSF2-206-exon2;SRSF2-novelTr003-exon2        1.0000000000 0.9917355372
#> SE:TP53-206-exon10;TP53-201-int9                   0.1145038168 0.0375939850
#> SE:TP53-201-exon10;TP53-217-int1                   1.0000000000 1.0000000000
#> SE:TP53-204-exon10;TP53-201-int9                   0.1830985915 0.0724637681
#> A5SS:SRSF2-novelTr001-exon1;SRSF2-209-exon1        0.1046153846 0.0567823344
#> TP53/ENST00000269305_Intron2/known-exon            0.0200759604 0.0076335878
#> A5SS:TRA2A-202-exon3;TRA2A-novelTr004-exon1        0.9411764706 1.0000000000
#> SE:SRSF2-novelTr004-exon2;SRSF2-209-int1           0.0820189274 0.0339256866
#> A5SS:TRA2A-208-exon2;TRA2A-novelTr003-exon1        1.0000000000 0.9722222222
#> SE:SRSF3-202-exon5;SRSF3-novelTr004-int1           0.9995078740 1.0000000000
#> A3SS:SRSF3-201-exon4;SRSF3-novelTr002-exon2        1.0000000000 0.9836065574
#> SE:SRSF1-201-exon2;SRSF1-novelTr004-int1           1.0000000000 0.9941520468
#> A5SS:SRSF1-novelTr003-exon1;SRSF1-201-exon3        0.0161725067 0.0054054054
#> SE:TP53-207-exon2;TP53-201-int1                    0.0000000000 0.0000000000
#> AFE:TP53-211-exon1;TP53-201-exon1                  0.0000000000 0.0000000000
#> SRSF2/ENST00000585202_Intron3/known-exon           0.8585653092 0.8197518676
#> SE:TRA2B-205-exon7;TRA2B-201-int3                  0.9961636829 0.9961795606
#> TRA2A/ENST00000297071_Intron7/clean                0.0799131445 0.0543219420
#> SE:TRA2B-206-exon2;TRA2B-205-int1                  0.6262686567 0.5135135135
#> AFE:TRA2B-211-exon1;TRA2B-205-exon1                0.6475225225 0.5280898876
#> SE:SRSF3-203-exon4;SRSF3-202-int3                  0.1420382166 0.1028255170
#> A3SS:SRSF1-204-exon4;SRSF1-203-exon2               0.9703703704 0.9728260870
#> SRSF1/ENST00000581979_Intron3/known-exon           0.4033282640 0.5138958951
#> SE:TRA2A-208-exon2;TRA2A-201-int1                  0.2158590308 0.1167728238
#> A5SS:TRA2A-208-exon2;TRA2A-203-exon2               0.8974358974 0.8974358974
#> A3SS:TRA2A-201-exon7;TRA2A-202-exon9               0.9766666667 0.9874371859
#> SE:TRA2A-202-exon3;TRA2A-212-int2                  0.6727272727 0.6190476190
#> SRSF1/ENST00000258962_Intron3/known-exon           0.2001967861 0.1499816371
#> A3SS:SRSF2-novelTr002-exon2;SRSF2-209-exon2        0.0582524272 0.0099337748
#> A5SS:SRSF1-208-exon4;SRSF1-novelTr001-exon1        0.1044776119 0.1915887850
#> SE:TRA2B-214-exon3;TRA2B-novelTr004-int1           0.9899772210 0.9968379447
#> SE:TRA2A-203-exon2;TRA2A-201-int1                  0.1583924350 0.0867178924
#> SE:TRA2B-novelTr015-exon2;TRA2B-214-int2           0.0170940171 0.0225303293
#> SE:TRA2A-202-exon2;TRA2A-208-int1                  0.1000000000 0.1329479769
#> NSUN5/ENST00000438747_Intron9/known-exon           0.3160804150 0.1556091677
#> MXE:TRA2A-212-exon2;TRA2A-211-exon2                0.1684210526 0.2222222222
#> A3SS:TP53-novelTr001-exon2;TP53-201-exon8          0.0000000000 0.0000000000
#> SE:TRA2A-211-exon2;TRA2A-201-int1                  0.1816091954 0.1053763441
#> A3SS:NSUN5-204-exon10;NSUN5-202-exon10             0.9680851064 0.9534883721
#> A5SS:TRA2A-208-exon2;TRA2A-novelTr004-exon1        0.9722222222 1.0000000000
#> SRSF3/ENST00000373715_Intron4/clean                0.0242105370 0.0086870469
#> A5SS:TRA2A-202-exon3;TRA2A-203-exon2               0.8000000000 0.8518518519
#> A5SS:SRSF1-207-exon3;SRSF1-novelTr001-exon1        0.0425531915 0.0225988701
#> A5SS:SRSF1-novelTr001-exon1;SRSF1-novelTr002-exon1 1.0000000000 0.9829545455
#> MXE:TRA2A-212-exon2;TRA2A-203-exon2                0.1927710843 0.2616822430
#> A3SS:TRA2A-201-exon3;TRA2A-novelTr008-exon2        0.9966367713 0.9962871287
#> SRSF3/ENST00000373715_Intron5/clean                0.0298916741 0.0161140469
#> A3SS:SRSF2-206-exon2;SRSF2-202-exon3               0.2108108108 0.2467443454
#> SE:SRSF3-novelTr008-exon2;SRSF3-202-int2           0.0009143554 0.0011827321
#> MXE:TRA2A-212-exon2;TRA2A-208-exon2                0.1403508772 0.2028985507
#> A5SS:SRSF3-202-exon1;SRSF3-novelTr005-exon1        1.0000000000 0.9995066601
#> A3SS:SRSF2-202-exon2;SRSF2-201-exon2               0.9883381924 0.9855072464
#> SE:SRSF2-206-exon2;SRSF2-202-int2                  0.2409669873 0.2900516796
#> MXE:SRSF1-209-exon2;SRSF1-novelTr005-exon2         1.0000000000 0.9863760218
#> A5SS:SRSF1-208-exon4;SRSF1-novelTr002-exon1        1.0000000000 0.9318181818
#> A3SS:NSUN5-201-exon8;NSUN5-novelTr003-exon2        0.8953488372 0.9633027523
#> SRSF2/ENST00000359995_Intron2/known-exon           0.1908173551 0.1122377562
#> A3SS:TRA2B-214-exon3;TRA2B-novelTr010-exon2        0.9930915371 0.9964664311
#> A5SS:SRSF2-206-exon2;SRSF2-202-exon2               0.2689033550 0.3286499695
#> ALE:TP53-210-exon5;TP53-216-exon2                  1.0000000000 1.0000000000
#> A3SS:NSUN5-201-exon7;NSUN5-206-exon6               0.7682926829 0.8991596639
#> A3SS:TRA2B-205-exon8;TRA2B-204-exon4               0.9833808167 0.9770609319
#> SE:TP53-201-exon6;TP53-216-int1                    1.0000000000 1.0000000000
#> AFE:SRSF3-208-exon1;SRSF3-202-exon1                0.0000000000 0.0000000000
#> AFE:TRA2A-204-exon1;TRA2A-201-exon1                0.0000000000 0.0000000000
#>                                                          02H026      02H033
#> AFE:TRA2B-215-exon1;TRA2B-203-exon1                0.9741518578 0.987485779
#> SE:TRA2B-205-exon2;TRA2B-203-int1                  0.9639504318 0.981915331
#> SE:TRA2B-novelTr014-exon2;TRA2B-214-int2           0.0058892815 0.053406998
#> SE:SRSF3-203-exon5;SRSF3-novelTr003-int1           0.9973663418 1.000000000
#> A5SS:SRSF2-202-exon1;SRSF2-205-exon1               0.9968652038 0.998175182
#> SE:TRA2B-205-exon4;TRA2B-novelTr003-int1           0.9849657307 0.993849308
#> A3SS:SRSF2-novelTr003-exon2;SRSF2-202-exon3        0.0011206575 0.003100089
#> SE:TRA2A-201-exon6;TRA2A-novelTr001-int1           0.8662519440 0.983471074
#> SE:NSUN5-206-exon5;NSUN5-novelTr001-int1           0.9740259740 0.996509599
#> SE:NSUN5-201-exon6;NSUN5-novelTr002-int1           0.9958762887 0.997780244
#> SRSF1/ENST00000584773_Intron4/known-exon           0.9333913569 0.938272053
#> SE:TRA2B-208-exon3;TRA2B-novelTr003-int1           0.9744552968 0.989271346
#> A3SS:SRSF1-201-exon4;SRSF1-novelTr001-exon2        0.9472759227 0.943548387
#> SE:SRSF1-207-exon3;SRSF1-novelTr001-int1           0.8999165972 0.893211289
#> A5SS:TRA2A-novelTr003-exon1;TRA2A-202-exon3        0.0000000000 0.142857143
#> SE:SRSF1-208-exon4;SRSF1-novelTr001-int1           0.9004975124 0.893778452
#> A5SS:SRSF1-206-exon1;SRSF1-201-exon1               0.0000000000 0.006305170
#> A5SS:SRSF2-211-exon1;SRSF2-202-exon2               0.0099962977 0.010114336
#> A5SS:TRA2A-208-exon2;TRA2A-202-exon3               0.7173913043 0.828571429
#> TRA2B/ENST00000453386_Intron4/clean                0.0228873504 0.027839669
#> A5SS:SRSF2-211-exon1;SRSF2-206-exon2               0.1516853933 0.060209424
#> SE:SRSF3-204-exon3;SRSF3-202-int2                  0.0000000000 0.000000000
#> SE:TRA2A-201-exon4;TRA2A-novelTr002-int1           0.9782608696 0.986733002
#> SE:TRA2A-212-exon2;TRA2A-201-int1                  0.0385259631 0.073732719
#> AFE:TRA2B-214-exon1;TRA2B-202-exon1                0.0332225914 0.008968610
#> A3SS:TRA2A-201-exon6;TRA2A-novelTr006-exon2        1.0000000000 0.962962963
#> SE:TP53-201-exon4;TP53-214-int3                    1.0000000000 1.000000000
#> A3SS:TP53-201-exon5;TP53-218-exon5                 1.0000000000 1.000000000
#> A3SS:TP53-201-exon2;TP53-205-exon2                 1.0000000000 0.994923858
#> NSUN5/ENST00000252594_Intron2/clean                0.0360809928 0.038090376
#> A3SS:TRA2B-208-exon3;TRA2B-205-exon4               0.0021401819 0.000000000
#> A5SS:TRA2B-214-exon2;TRA2B-novelTr008-exon1        1.0000000000 0.996124031
#> A3SS:NSUN5-201-exon3;NSUN5-203-exon3               1.0000000000 0.993527508
#> A5SS:SRSF3-202-exon5;SRSF3-novelTr007-exon1        0.9995402299 0.999161777
#> A3SS:SRSF2-206-exon2;SRSF2-novelTr003-exon2        0.9727272727 0.973977695
#> SE:TP53-206-exon10;TP53-201-int9                   0.0237717908 0.007547170
#> SE:TP53-201-exon10;TP53-217-int1                   1.0000000000 1.000000000
#> SE:TP53-204-exon10;TP53-201-int9                   0.0344827586 0.021395349
#> A5SS:SRSF2-novelTr001-exon1;SRSF2-209-exon1        0.2528735632 0.181818182
#> TP53/ENST00000269305_Intron2/known-exon            0.0022295624 0.000000000
#> A5SS:TRA2A-202-exon3;TRA2A-novelTr004-exon1        1.0000000000 0.923076923
#> SE:SRSF2-novelTr004-exon2;SRSF2-209-int1           0.1798107256 0.157894737
#> A5SS:TRA2A-208-exon2;TRA2A-novelTr003-exon1        1.0000000000 0.966666667
#> SE:SRSF3-202-exon5;SRSF3-novelTr004-int1           0.9997497811 0.999420794
#> A3SS:SRSF3-201-exon4;SRSF3-novelTr002-exon2        0.9871794872 0.980000000
#> SE:SRSF1-201-exon2;SRSF1-novelTr004-int1           0.9983539095 1.000000000
#> A5SS:SRSF1-novelTr003-exon1;SRSF1-201-exon3        0.0009267841 0.002557545
#> SE:TP53-207-exon2;TP53-201-int1                    0.0000000000 0.000000000
#> AFE:TP53-211-exon1;TP53-201-exon1                  0.0000000000 0.000000000
#> SRSF2/ENST00000585202_Intron3/known-exon           0.9226409631 0.952327911
#> SE:TRA2B-205-exon7;TRA2B-201-int3                  0.9923371648 0.987970877
#> TRA2A/ENST00000297071_Intron7/clean                0.0246483052 0.058494476
#> SE:TRA2B-206-exon2;TRA2B-205-int1                  0.3198743831 0.267937220
#> AFE:TRA2B-211-exon1;TRA2B-205-exon1                0.3576271186 0.282417582
#> SE:SRSF3-203-exon4;SRSF3-202-int3                  0.0301295160 0.043893424
#> A3SS:SRSF1-204-exon4;SRSF1-203-exon2               0.9466666667 0.968911917
#> SRSF1/ENST00000581979_Intron3/known-exon           0.9362123384 0.861177328
#> SE:TRA2A-208-exon2;TRA2A-201-int1                  0.1196319018 0.275675676
#> A5SS:TRA2A-208-exon2;TRA2A-203-exon2               0.8461538462 0.852941176
#> A3SS:TRA2A-201-exon7;TRA2A-202-exon9               0.9702602230 0.965831435
#> SE:TRA2A-202-exon3;TRA2A-212-int2                  0.4146341463 0.512195122
#> SRSF1/ENST00000258962_Intron3/known-exon           0.0319601935 0.079567125
#> A3SS:SRSF2-novelTr002-exon2;SRSF2-209-exon2        0.0909090909 0.132530120
#> A5SS:SRSF1-208-exon4;SRSF1-novelTr001-exon1        0.1176470588 0.102564103
#> SE:TRA2B-214-exon3;TRA2B-novelTr004-int1           0.9946500223 0.996001999
#> SE:TRA2A-203-exon2;TRA2A-201-int1                  0.0816000000 0.207100592
#> SE:TRA2B-novelTr015-exon2;TRA2B-214-int2           0.0332187858 0.033834586
#> SE:TRA2A-202-exon2;TRA2A-208-int1                  0.1428571429 0.140271493
#> NSUN5/ENST00000438747_Intron9/known-exon           0.0417164432 0.096411714
#> MXE:TRA2A-212-exon2;TRA2A-211-exon2                0.2839506173 0.230215827
#> A3SS:TP53-novelTr001-exon2;TP53-201-exon8          0.0021141649 0.002547771
#> SE:TRA2A-211-exon2;TRA2A-201-int1                  0.0917721519 0.210216110
#> A3SS:NSUN5-204-exon10;NSUN5-202-exon10             0.9433962264 0.939914163
#> A5SS:TRA2A-208-exon2;TRA2A-novelTr004-exon1        1.0000000000 0.983050847
#> SRSF3/ENST00000373715_Intron4/clean                0.0049179952 0.007598906
#> A5SS:TRA2A-202-exon3;TRA2A-203-exon2               0.6842105263 0.545454545
#> A5SS:SRSF1-207-exon3;SRSF1-novelTr001-exon1        0.0163934426 0.014084507
#> A5SS:SRSF1-novelTr001-exon1;SRSF1-novelTr002-exon1 0.9375000000 0.972222222
#> MXE:TRA2A-212-exon2;TRA2A-203-exon2                0.3108108108 0.233576642
#> A3SS:TRA2A-201-exon3;TRA2A-novelTr008-exon2        0.9945553539 0.996226415
#> SRSF3/ENST00000373715_Intron5/clean                0.0034718619 0.011325239
#> A3SS:SRSF2-206-exon2;SRSF2-202-exon3               0.0384753686 0.104257859
#> SE:SRSF3-novelTr008-exon2;SRSF3-202-int2           0.0000000000 0.000000000
#> MXE:TRA2A-212-exon2;TRA2A-208-exon2                0.2277227723 0.172972973
#> A5SS:SRSF3-202-exon1;SRSF3-novelTr005-exon1        0.9995478182 0.999741602
#> A3SS:SRSF2-202-exon2;SRSF2-201-exon2               0.9906542056 0.980286738
#> SE:SRSF2-206-exon2;SRSF2-202-int2                  0.0460221192 0.121218036
#> MXE:SRSF1-209-exon2;SRSF1-novelTr005-exon2         0.9919614148 0.997468354
#> A5SS:SRSF1-208-exon4;SRSF1-novelTr002-exon1        0.6666666667 0.800000000
#> A3SS:NSUN5-201-exon8;NSUN5-novelTr003-exon2        1.0000000000 0.961904762
#> SRSF2/ENST00000359995_Intron2/known-exon           0.0576131373 0.084019372
#> A3SS:TRA2B-214-exon3;TRA2B-novelTr010-exon2        1.0000000000 0.992277992
#> A5SS:SRSF2-206-exon2;SRSF2-202-exon2               0.0534513274 0.137547893
#> ALE:TP53-210-exon5;TP53-216-exon2                  0.9965753425 1.000000000
#> A3SS:NSUN5-201-exon7;NSUN5-206-exon6               0.9692307692 0.903940887
#> A3SS:TRA2B-205-exon8;TRA2B-204-exon4               0.9803843074 0.986473430
#> SE:TP53-201-exon6;TP53-216-int1                    0.9968895801 1.000000000
#> AFE:SRSF3-208-exon1;SRSF3-202-exon1                0.0000000000 0.000000000
#> AFE:TRA2A-204-exon1;TRA2A-201-exon1                0.0000000000 0.000000000
#>                                                          02H043       02H046
#> AFE:TRA2B-215-exon1;TRA2B-203-exon1                0.9869822485 0.9888268156
#> SE:TRA2B-205-exon2;TRA2B-203-int1                  0.9796484736 0.9835514019
#> SE:TRA2B-novelTr014-exon2;TRA2B-214-int2           0.0018214936 0.0027598896
#> SE:SRSF3-203-exon5;SRSF3-novelTr003-int1           1.0000000000 0.9985465116
#> A5SS:SRSF2-202-exon1;SRSF2-205-exon1               1.0000000000 1.0000000000
#> SE:TRA2B-205-exon4;TRA2B-novelTr003-int1           1.0000000000 0.9887926643
#> A3SS:SRSF2-novelTr003-exon2;SRSF2-202-exon3        0.0000000000 0.0067873303
#> SE:TRA2A-201-exon6;TRA2A-novelTr001-int1           0.9813519814 0.9708994709
#> SE:NSUN5-206-exon5;NSUN5-novelTr001-int1           0.9925373134 0.9851851852
#> SE:NSUN5-201-exon6;NSUN5-novelTr002-int1           0.9951100244 1.0000000000
#> SRSF1/ENST00000584773_Intron4/known-exon           0.8552751581 0.8767239216
#> SE:TRA2B-208-exon3;TRA2B-novelTr003-int1           1.0000000000 0.9796860572
#> A3SS:SRSF1-201-exon4;SRSF1-novelTr001-exon2        0.7952755906 0.8934624697
#> SE:SRSF1-207-exon3;SRSF1-novelTr001-int1           0.6630669546 0.8076502732
#> A5SS:TRA2A-novelTr003-exon1;TRA2A-202-exon3        0.0833333333 0.0000000000
#> SE:SRSF1-208-exon4;SRSF1-novelTr001-int1           0.6822810591 0.8103448276
#> A5SS:SRSF1-206-exon1;SRSF1-201-exon1               0.0000000000 0.0053571429
#> A5SS:SRSF2-211-exon1;SRSF2-202-exon2               0.0044085231 0.0407865987
#> A5SS:TRA2A-208-exon2;TRA2A-202-exon3               0.6666666667 0.8023255814
#> TRA2B/ENST00000453386_Intron4/clean                0.0261475462 0.0392895743
#> A5SS:SRSF2-211-exon1;SRSF2-206-exon2               0.0164835165 0.1076923077
#> SE:SRSF3-204-exon3;SRSF3-202-int2                  0.0007265682 0.0005821851
#> SE:TRA2A-201-exon4;TRA2A-novelTr002-int1           0.9891774892 0.9874213836
#> SE:TRA2A-212-exon2;TRA2A-201-int1                  0.0417827298 0.0550335570
#> AFE:TRA2B-214-exon1;TRA2B-202-exon1                0.0060240964 0.0272727273
#> A3SS:TRA2A-201-exon6;TRA2A-novelTr006-exon2        0.9473684211 1.0000000000
#> SE:TP53-201-exon4;TP53-214-int3                    1.0000000000 1.0000000000
#> A3SS:TP53-201-exon5;TP53-218-exon5                 1.0000000000 1.0000000000
#> A3SS:TP53-201-exon2;TP53-205-exon2                 0.9930555556 0.9757085020
#> NSUN5/ENST00000252594_Intron2/clean                0.0368718661 0.0431391259
#> A3SS:TRA2B-208-exon3;TRA2B-205-exon4               0.0012919897 0.0000000000
#> A5SS:TRA2B-214-exon2;TRA2B-novelTr008-exon1        0.9963636364 0.9981583794
#> A3SS:NSUN5-201-exon3;NSUN5-203-exon3               1.0000000000 0.9945945946
#> A5SS:SRSF3-202-exon5;SRSF3-novelTr007-exon1        0.9986842105 0.9993048314
#> A3SS:SRSF2-206-exon2;SRSF2-novelTr003-exon2        1.0000000000 0.9700000000
#> SE:TP53-206-exon10;TP53-201-int9                   0.0190735695 0.0328358209
#> SE:TP53-201-exon10;TP53-217-int1                   0.9898477157 1.0000000000
#> SE:TP53-204-exon10;TP53-201-int9                   0.0270270270 0.0540145985
#> A5SS:SRSF2-novelTr001-exon1;SRSF2-209-exon1        0.0097799511 0.0985010707
#> TP53/ENST00000269305_Intron2/known-exon            0.0051913000 0.0055441256
#> A5SS:TRA2A-202-exon3;TRA2A-novelTr004-exon1        1.0000000000 0.8947368421
#> SE:SRSF2-novelTr004-exon2;SRSF2-209-int1           0.0085679315 0.0696132597
#> A5SS:TRA2A-208-exon2;TRA2A-novelTr003-exon1        0.9565217391 1.0000000000
#> SE:SRSF3-202-exon5;SRSF3-novelTr004-int1           1.0000000000 0.9992411307
#> A3SS:SRSF3-201-exon4;SRSF3-novelTr002-exon2        1.0000000000 0.9606741573
#> SE:SRSF1-201-exon2;SRSF1-novelTr004-int1           1.0000000000 0.9866412214
#> A5SS:SRSF1-novelTr003-exon1;SRSF1-201-exon3        0.0000000000 0.0053908356
#> SE:TP53-207-exon2;TP53-201-int1                    0.0000000000 0.0000000000
#> AFE:TP53-211-exon1;TP53-201-exon1                  0.0000000000 0.0000000000
#> SRSF2/ENST00000585202_Intron3/known-exon           0.7538079104 0.7271061557
#> SE:TRA2B-205-exon7;TRA2B-201-int3                  0.9969581749 0.9941137999
#> TRA2A/ENST00000297071_Intron7/clean                0.1154078082 0.0756229889
#> SE:TRA2B-206-exon2;TRA2B-205-int1                  0.4938132733 0.4148845167
#> AFE:TRA2B-211-exon1;TRA2B-205-exon1                0.5490981964 0.4420880914
#> SE:SRSF3-203-exon4;SRSF3-202-int3                  0.1610805403 0.1137491906
#> A3SS:SRSF1-204-exon4;SRSF1-203-exon2               0.9958071279 0.7829457364
#> SRSF1/ENST00000581979_Intron3/known-exon           0.4443609627 0.8447012398
#> SE:TRA2A-208-exon2;TRA2A-201-int1                  0.1421446384 0.1880046136
#> A5SS:TRA2A-208-exon2;TRA2A-203-exon2               1.0000000000 0.9583333333
#> A3SS:TRA2A-201-exon7;TRA2A-202-exon9               0.9757869249 0.9719101124
#> SE:TRA2A-202-exon3;TRA2A-212-int2                  0.5000000000 0.4716981132
#> SRSF1/ENST00000258962_Intron3/known-exon           0.3597994150 0.0810912343
#> A3SS:SRSF2-novelTr002-exon2;SRSF2-209-exon2        0.0073529412 0.0388127854
#> A5SS:SRSF1-208-exon4;SRSF1-novelTr001-exon1        0.2909090909 0.1372549020
#> SE:TRA2B-214-exon3;TRA2B-novelTr004-int1           0.9946140036 0.9944067119
#> SE:TRA2A-203-exon2;TRA2A-201-int1                  0.0923482850 0.1210986267
#> SE:TRA2B-novelTr015-exon2;TRA2B-214-int2           0.0000000000 0.0198915009
#> SE:TRA2A-202-exon2;TRA2A-208-int1                  0.1250000000 0.1569506726
#> NSUN5/ENST00000438747_Intron9/known-exon           0.1493838183 0.1193003272
#> MXE:TRA2A-212-exon2;TRA2A-211-exon2                0.2459016393 0.2697368421
#> A3SS:TP53-novelTr001-exon2;TP53-201-exon8          0.0041666667 0.0000000000
#> SE:TRA2A-211-exon2;TRA2A-201-int1                  0.1179487179 0.1361963190
#> A3SS:NSUN5-204-exon10;NSUN5-202-exon10             0.9606299213 0.9220779221
#> A5SS:TRA2A-208-exon2;TRA2A-novelTr004-exon1        1.0000000000 0.9718309859
#> SRSF3/ENST00000373715_Intron4/clean                0.0105728480 0.0202336061
#> A5SS:TRA2A-202-exon3;TRA2A-203-exon2               1.0000000000 0.8500000000
#> A5SS:SRSF1-207-exon3;SRSF1-novelTr001-exon1        0.0487804878 0.0112359551
#> A5SS:SRSF1-novelTr001-exon1;SRSF1-novelTr002-exon1 0.9750000000 0.9887640449
#> MXE:TRA2A-212-exon2;TRA2A-203-exon2                0.3000000000 0.2971014493
#> A3SS:TRA2A-201-exon3;TRA2A-novelTr008-exon2        0.9927007299 0.9958620690
#> SRSF3/ENST00000373715_Intron5/clean                0.0107187174 0.0149727899
#> A3SS:SRSF2-206-exon2;SRSF2-202-exon3               0.1905615293 0.1809701493
#> SE:SRSF3-novelTr008-exon2;SRSF3-202-int2           0.0004844961 0.0009699321
#> MXE:TRA2A-212-exon2;TRA2A-208-exon2                0.2083333333 0.2009803922
#> A5SS:SRSF3-202-exon1;SRSF3-novelTr005-exon1        1.0000000000 0.9993614304
#> A3SS:SRSF2-202-exon2;SRSF2-201-exon2               0.9901315789 0.9952324195
#> SE:SRSF2-206-exon2;SRSF2-202-int2                  0.1998819014 0.2227795810
#> MXE:SRSF1-209-exon2;SRSF1-novelTr005-exon2         0.9947229551 0.9858407080
#> A5SS:SRSF1-208-exon4;SRSF1-novelTr002-exon1        0.9411764706 0.9333333333
#> A3SS:NSUN5-201-exon8;NSUN5-novelTr003-exon2        0.9950495050 0.9177215190
#> SRSF2/ENST00000359995_Intron2/known-exon           0.1488402008 0.1468749763
#> A3SS:TRA2B-214-exon3;TRA2B-novelTr010-exon2        1.0000000000 0.9963235294
#> A5SS:SRSF2-206-exon2;SRSF2-202-exon2               0.2089900759 0.2605277934
#> ALE:TP53-210-exon5;TP53-216-exon2                  1.0000000000 0.9927797834
#> A3SS:NSUN5-201-exon7;NSUN5-206-exon6               0.9221556886 0.8972602740
#> A3SS:TRA2B-205-exon8;TRA2B-204-exon4               0.9840000000 0.9737458978
#> SE:TP53-201-exon6;TP53-216-int1                    1.0000000000 0.9930555556
#> AFE:SRSF3-208-exon1;SRSF3-202-exon1                0.0000000000 0.0000000000
#> AFE:TRA2A-204-exon1;TRA2A-201-exon1                0.0000000000 0.0000000000