Perform over representation analysis (ORA) against Molecular Signature Database (MsigDB) .

Perform over representation analysis (ORA) against Molecular Signature Database (MsigDB) .

Usage

enrichMsigDB(
  gene_list,
  from_type = "SYMBOL",
  orgdb = org.Mm.eg.db::org.Mm.eg.db,
  msigdb_category = "H",
  species = "Mus musculus",
  background = NULL,
  col_genetype = "GENETYPE",
  feature_type = "protein-coding",
  ...
)

Arguments

gene_list

a character vector denoting valid gene names.

from_type

a string, default SYMBOL, denoting the type of input gene_list (e.g SYMBOL, ENTREZID).

orgdb

an object of the class OrgDB. Default org.Mm.eg.db::org.Mm.eg.db.

msigdb_category

a character string, default "H", denoting one of the nine catagories of geneList given in the msigdb. Possible values are: "H","C1","C2","C3","C4","C5","C6","C7", and "C8".

species

a character string denoting name of the species. Default "Mus musculus". Possible values can be found via the function msigdbr::msigdbr_species().

background

NULL (default) or a character vector. If NULL, all features filtered by feature_type will be selected as background. In case of character vector, values will be used as it is for background set. In case of user provided background set it is assumed that id type is same as queried genes_list (i.e. from_type). NOTE: In both scenarios values obtained from MSigDB catagory database will always remain as a part of background set.

col_genetype

A string denoting a column name to filter feature_type.Valid only if the argument background is NULL.

feature_type

A string denoting a value to filter the column mentioned by the argument col_genetype.Valid only if the argument background is NULL.

...

Other parameters to pass clusterProfiler::enricher().

Value

Enrichment results.

Examples

if (FALSE) {

}