Get Alternate Splice Events based on their differential regulation.

Get Alternate Splice Events based on their differential regulation.

Usage

get_ASE_by_regulation(
  x,
  sample_comparisons,
  regulation = "both",
  simplify = FALSE
)

get_ASEsets_by_regulation(x, sample_comparisons, regulation = "both")

Arguments

x

an abject of class "parcutils_ase". This is an output of the function run_ase_diff_analysis().

sample_comparisons

a character vector denoting sample comparisons for which ASE to be obtained.

regulation

a character string, default both. Values can be one of the up, down, both, other, all.

up : returns all up regulated ASE.

down : returns all down regulated ASE

both : returns all up and down regulated ASE.

other : returns ASE other than up and down regulated ASE.

all : returns all ASE.

simplify

logical, default FALSE, if TRUE returns result in a dataframe format.

Value

a list or dataframe.

Details

For a given differential comparison this function returns up, down, both, other and all ASE.

Examples

se <- SpliceWiz::SpliceWiz_example_NxtSE(novelSplicing = TRUE)
SpliceWiz::colData(se)$treatment <- rep(c("A", "B"), each = 3)
SpliceWiz::colData(se)$replicate <- rep(c("P","Q","R"), 2)

res <- run_splicewiz_diff_analysis(x = se, test_factor = "treatment", test_nom = "A" ,test_denom = "B",  IRmode ="annotated")
#> Error in run_splicewiz_diff_analysis(x = se, test_factor = "treatment",     test_nom = "A", test_denom = "B", IRmode = "annotated"): could not find function "run_splicewiz_diff_analysis"

# get both up and down regulated genes
get_ASE_by_regulation(x = res, sample_comparisons = c("A_VS_B")) %>% head()
#>      AFE:TRA2B-215-exon1;TRA2B-203-exon1 
#>                                     down 
#>        SE:TRA2B-205-exon2;TRA2B-203-int1 
#>                                     down 
#> SE:TRA2B-novelTr014-exon2;TRA2B-214-int2 
#>                                     down 
#> SE:SRSF3-203-exon5;SRSF3-novelTr003-int1 
#>                                     down 
#>     A5SS:SRSF2-202-exon1;SRSF2-205-exon1 
#>                                     down 
#> SE:TRA2B-205-exon4;TRA2B-novelTr003-int1 
#>                                     down 
#> Levels: up down other

# get up genes only
get_ASE_by_regulation(x = res, sample_comparisons = c("A_VS_B") , regul = "up") %>% head()
#>                A3SS:TP53-201-exon2;TP53-205-exon2 
#>                                                up 
#>               NSUN5/ENST00000252594_Intron2/clean 
#>                                                up 
#> MXE:TRA2B-novelTr015-exon2;TRA2B-novelTr014-exon2 
#>                                                up 
#>              A3SS:TRA2B-208-exon3;TRA2B-205-exon4 
#>                                                up 
#>       A5SS:TRA2B-214-exon2;TRA2B-novelTr008-exon1 
#>                                                up 
#>              A3SS:SRSF2-205-exon2;SRSF2-207-exon2 
#>                                                up 
#> Levels: up down other

# get down genes only
get_ASE_by_regulation(x = res, sample_comparisons = c("A_VS_B") , regul = "down") %>% head()
#>      AFE:TRA2B-215-exon1;TRA2B-203-exon1 
#>                                     down 
#>        SE:TRA2B-205-exon2;TRA2B-203-int1 
#>                                     down 
#> SE:TRA2B-novelTr014-exon2;TRA2B-214-int2 
#>                                     down 
#> SE:SRSF3-203-exon5;SRSF3-novelTr003-int1 
#>                                     down 
#>     A5SS:SRSF2-202-exon1;SRSF2-205-exon1 
#>                                     down 
#> SE:TRA2B-205-exon4;TRA2B-novelTr003-int1 
#>                                     down 
#> Levels: up down other

# get genes other than up and down
get_ASE_by_regulation(x = res, sample_comparisons = c("A_VS_B") , regul = "other") %>% head()
#> SRSF2/ENST00000585202_Intron3/known-exon 
#>                                    other 
#>        SE:TRA2B-205-exon7;TRA2B-201-int3 
#>                                    other 
#>      TRA2A/ENST00000297071_Intron7/clean 
#>                                    other 
#>        SE:TRA2B-206-exon2;TRA2B-205-int1 
#>                                    other 
#>      AFE:TRA2B-211-exon1;TRA2B-205-exon1 
#>                                    other 
#>        SE:SRSF3-203-exon4;SRSF3-202-int3 
#>                                    other 
#> Levels: up down other

# Simplify output for multiple sample comparisons
get_ASE_by_regulation(x = res, sample_comparisons = res$de_comparisons, simplify = TRUE, regul= "up")
#> # A tibble: 27 × 3
#>    sample_comparisons ase                                               regul
#>    <chr>              <chr>                                             <fct>
#>  1 A_VS_B             A3SS:TP53-201-exon2;TP53-205-exon2                up   
#>  2 A_VS_B             NSUN5/ENST00000252594_Intron2/clean               up   
#>  3 A_VS_B             MXE:TRA2B-novelTr015-exon2;TRA2B-novelTr014-exon2 up   
#>  4 A_VS_B             A3SS:TRA2B-208-exon3;TRA2B-205-exon4              up   
#>  5 A_VS_B             A5SS:TRA2B-214-exon2;TRA2B-novelTr008-exon1       up   
#>  6 A_VS_B             A3SS:SRSF2-205-exon2;SRSF2-207-exon2              up   
#>  7 A_VS_B             A3SS:NSUN5-201-exon3;NSUN5-203-exon3              up   
#>  8 A_VS_B             A5SS:SRSF3-202-exon5;SRSF3-novelTr007-exon1       up   
#>  9 A_VS_B             A3SS:SRSF2-206-exon2;SRSF2-novelTr003-exon2       up   
#> 10 A_VS_B             SE:TP53-206-exon10;TP53-201-int9                  up   
#> # … with 17 more rows

# get genesets by regulation. It uses sample comparison and regulation to name each output geneset.

get_ASEsets_by_regulation(x = res, sample_comparisons = "A_VS_B", regul = "both")
#> $A_VS_B_down
#>  [1] "AFE:TRA2B-215-exon1;TRA2B-203-exon1"               
#>  [2] "SE:TRA2B-205-exon2;TRA2B-203-int1"                 
#>  [3] "SE:TRA2B-novelTr014-exon2;TRA2B-214-int2"          
#>  [4] "SE:SRSF3-203-exon5;SRSF3-novelTr003-int1"          
#>  [5] "A5SS:SRSF2-202-exon1;SRSF2-205-exon1"              
#>  [6] "SE:TRA2B-205-exon4;TRA2B-novelTr003-int1"          
#>  [7] "A3SS:SRSF2-novelTr003-exon2;SRSF2-202-exon3"       
#>  [8] "SE:TRA2A-201-exon6;TRA2A-novelTr001-int1"          
#>  [9] "SE:NSUN5-206-exon5;NSUN5-novelTr001-int1"          
#> [10] "SE:NSUN5-201-exon6;NSUN5-novelTr002-int1"          
#> [11] "SRSF1/ENST00000584773_Intron4/known-exon"          
#> [12] "SE:TRA2B-208-exon3;TRA2B-novelTr003-int1"          
#> [13] "A3SS:TRA2A-novelTr001-exon2;TRA2A-novelTr005-exon2"
#> [14] "A3SS:SRSF1-201-exon4;SRSF1-novelTr001-exon2"       
#> [15] "SE:SRSF1-207-exon3;SRSF1-novelTr001-int1"          
#> [16] "A5SS:TRA2A-novelTr003-exon1;TRA2A-202-exon3"       
#> [17] "SE:SRSF1-208-exon4;SRSF1-novelTr001-int1"          
#> [18] "A5SS:SRSF1-206-exon1;SRSF1-201-exon1"              
#> [19] "A5SS:TRA2A-novelTr003-exon1;TRA2A-203-exon2"       
#> [20] "A5SS:SRSF2-211-exon1;SRSF2-202-exon2"              
#> [21] "A5SS:TRA2A-208-exon2;TRA2A-202-exon3"              
#> [22] "TRA2B/ENST00000453386_Intron4/clean"               
#> [23] "A5SS:SRSF2-211-exon1;SRSF2-206-exon2"              
#> [24] "SE:SRSF3-204-exon3;SRSF3-202-int2"                 
#> [25] "SE:TRA2A-201-exon4;TRA2A-novelTr002-int1"          
#> [26] "SE:TRA2A-212-exon2;TRA2A-201-int1"                 
#> [27] "AFE:TRA2B-214-exon1;TRA2B-202-exon1"               
#> [28] "A3SS:TRA2A-201-exon6;TRA2A-novelTr006-exon2"       
#> [29] "A5SS:TRA2A-novelTr004-exon1;TRA2A-203-exon2"       
#> [30] "A5SS:TRA2A-novelTr003-exon1;TRA2A-novelTr004-exon1"
#> [31] "ALE:TP53-217-exon2;TP53-202-exon9"                 
#> [32] "SE:TP53-201-exon4;TP53-214-int3"                   
#> [33] "A3SS:TP53-201-exon5;TP53-218-exon5"                
#> 
#> $A_VS_B_up
#>  [1] "A3SS:TP53-201-exon2;TP53-205-exon2"               
#>  [2] "NSUN5/ENST00000252594_Intron2/clean"              
#>  [3] "MXE:TRA2B-novelTr015-exon2;TRA2B-novelTr014-exon2"
#>  [4] "A3SS:TRA2B-208-exon3;TRA2B-205-exon4"             
#>  [5] "A5SS:TRA2B-214-exon2;TRA2B-novelTr008-exon1"      
#>  [6] "A3SS:SRSF2-205-exon2;SRSF2-207-exon2"             
#>  [7] "A3SS:NSUN5-201-exon3;NSUN5-203-exon3"             
#>  [8] "A5SS:SRSF3-202-exon5;SRSF3-novelTr007-exon1"      
#>  [9] "A3SS:SRSF2-206-exon2;SRSF2-novelTr003-exon2"      
#> [10] "SE:TP53-206-exon10;TP53-201-int9"                 
#> [11] "SE:TP53-201-exon10;TP53-217-int1"                 
#> [12] "SE:TP53-204-exon10;TP53-201-int9"                 
#> [13] "A5SS:SRSF2-201-exon1;SRSF2-207-exon1"             
#> [14] "A5SS:SRSF2-novelTr001-exon1;SRSF2-209-exon1"      
#> [15] "TP53/ENST00000269305_Intron2/known-exon"          
#> [16] "A5SS:TRA2A-202-exon3;TRA2A-novelTr004-exon1"      
#> [17] "SE:SRSF2-novelTr004-exon2;SRSF2-209-int1"         
#> [18] "A5SS:TRA2A-208-exon2;TRA2A-novelTr003-exon1"      
#> [19] "SE:SRSF3-202-exon5;SRSF3-novelTr004-int1"         
#> [20] "MXE:SRSF3-novelTr008-exon2;SRSF3-204-exon3"       
#> [21] "A5SS:SRSF1-207-exon3;SRSF1-208-exon4"             
#> [22] "A3SS:SRSF3-201-exon4;SRSF3-novelTr002-exon2"      
#> [23] "SE:SRSF1-201-exon2;SRSF1-novelTr004-int1"         
#> [24] "A5SS:SRSF1-novelTr003-exon1;SRSF1-201-exon3"      
#> [25] "A5SS:SRSF1-207-exon3;SRSF1-novelTr002-exon1"      
#> [26] "SE:TP53-207-exon2;TP53-201-int1"                  
#> [27] "AFE:TP53-211-exon1;TP53-201-exon1"                
#>