Get a data matrix (PSI, logit, z-score) for a given ASE

Get a data matrix (PSI, logit, z-score) for a given ASE

Usage

get_ASE_data_matrix(
  se,
  event_names,
  samples,
  replicates = TRUE,
  method = "PSI",
  column_condition = "condition"
)

Arguments

se

an object of class NxtSE.

event_names

a character vector denoting valid ASE names to plot in the heatmap.

samples

a character vector denoting valid sample names to plot in the heatmap.

replicates

logical, default TRUE, denoting whether to show sample replicates or not.

method

a character string denoting a method for data value. Can be one of the "PSI", "logit" or "Z-Score"

column_condtion

a name of column in se storing condition. Default: "condition"

Value

a data matrix

Examples

se <- SpliceWiz::SpliceWiz_example_NxtSE(novelSplicing = TRUE)
SpliceWiz::colData(se)$treatment <- rep(c("A", "B"), each = 3)
SpliceWiz::colData(se)$replicate <- rep(c("P","Q","R"), 2)
res <- run_ase_diff_analysis(x = se, test_factor = "treatment", test_nom = "A" ,test_denom = "B",  IRmode ="annotated",  cutoff_lfc = 0.6, cutoff_padj = 1, regul_based_upon = 2)
#> Mar 23 16:08:55 Performing edgeR contrast for included / excluded counts separately
#> Mar 23 16:09:03 Performing edgeR contrast for included / excluded counts together
event_names = get_ASEsets_by_regulation(x = res, sample_comparisons = "A_VS_B", regul = "all") %>% unlist()
get_ASE_data_matrix(se, event_names , samples = c("A", "B"), column_condition ="treatment")
#> # A tibble: 99 × 7
#>    event_name                    02H00…¹ 02H02…² 02H02…³ 02H03…⁴ 02H04…⁵ 02H04…⁶
#>    <chr>                           <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#>  1 AFE:TRA2B-215-exon1;TRA2B-20… 0.963   0.950   0.974   0.987   0.987   0.989  
#>  2 SE:TRA2B-205-exon2;TRA2B-203… 0.940   0.923   0.964   0.982   0.980   0.984  
#>  3 SE:TRA2B-novelTr014-exon2;TR… 0.00260 0       0.00589 0.0534  0.00182 0.00276
#>  4 SE:SRSF3-203-exon5;SRSF3-nov… 0.998   0.997   0.997   1       1       0.999  
#>  5 A5SS:SRSF2-202-exon1;SRSF2-2… 0.994   0.997   0.997   0.998   1       1      
#>  6 SE:TRA2B-205-exon4;TRA2B-nov… 0.977   0.992   0.985   0.994   1       0.989  
#>  7 A3SS:SRSF2-novelTr003-exon2;… 0       0.00272 0.00112 0.00310 0       0.00679
#>  8 SE:TRA2A-201-exon6;TRA2A-nov… 0.983   0.951   0.866   0.983   0.981   0.971  
#>  9 SE:NSUN5-206-exon5;NSUN5-nov… 0.969   0.980   0.974   0.997   0.993   0.985  
#> 10 SE:NSUN5-201-exon6;NSUN5-nov… 0.988   0.980   0.996   0.998   0.995   1      
#> # … with 89 more rows, and abbreviated variable names ¹​`02H003`, ²​`02H025`,
#> #   ³​`02H026`, ⁴​`02H033`, ⁵​`02H043`, ⁶​`02H046`