Generate a heatmap of normalised gene expression values, z-score or log2Fold-change values.

Heatmap is a common tool to show gene expression pattern across samples in RNA-seq experiment. While doing data exploration, it is a common practice to generate several heatmaps to identify interesting patterns of gene expression across sample. Most heatmap generating tools require data in a tabular form. However, to prepare such a tabular form requires significant data wrangling such as sub-setting genes (rows) and samples (columns) specifically when RNA-seq studies studies consisting of several samples and replicates.

This function cut downs several steps of data wrangling to create a heatmap of gene expression, z-score or log2 fold-change . Given an object of 'parcutils', names of samples or sample comprisons, genes to show in heatmap and several other arguments it creates a heatmap.The output of the function is an output of the function ComplexHeatmap::Heatmap() which then can be used for the other functions of the ComplexHeatmap package.

Usage

get_gene_expression_heatmap(
  x,
  samples,
  genes,
  repair_genes = FALSE,
  convert_log2 = FALSE,
  color_default = TRUE,
  col = NULL,
  convert_zscore = TRUE,
  summarise_replicates = TRUE,
  summarise_method = "median",
  show_row_names = FALSE,
  cluster_rows = TRUE,
  show_row_dend = TRUE,
  row_names_font_size = 10,
  show_column_names = TRUE,
  cluster_columns = TRUE,
  show_heatmap_legend = TRUE,
  ...
)

get_fold_change_heatmap(
  x,
  sample_comparisons,
  genes,
  repair_genes = FALSE,
  color_default = TRUE,
  col = NULL,
  show_row_names = FALSE,
  cluster_rows = TRUE,
  show_row_dend = TRUE,
  row_names_font_size = 10,
  show_column_names = TRUE,
  cluster_columns = TRUE,
  show_heatmap_legend = TRUE,
  ...
)

Arguments

x

an abject of class "parcutils". This is an output of the function run_deseq_analysis().

samples

a character vector denoting sample names to show in the heatmap.

genes

a character vector denoting genes to show in the heatmap.

repair_genes

logical, default FALSE, indicating whether to repair gene names or not. See details.

convert_log2

logical, default FALSE, indicating whether to convert gene expression values in log2 or not.

color_default

logical, default TRUE, indicating whether to use default heatmap colors.

col

an output of circlize::colorRamp2(), default NULL.

convert_zscore

logical, default TRUE, indicating whether to convert gene expression values in to the z-score or not. see details.

summarise_replicates

logical, default TRUE, indicating whether to summarise values for each gene across replicates.

summarise_method

a character string, either mean or median.

show_row_names

logical, default FALSE, indicating whether to show row names in the heatmap or not. Internally this argument is passed to the function ComplexHeatmap::Heatmap().

cluster_rows

logical, default TRUE, indicating whether to cluster rows in the heatmap or not. Internally this argument is passed to the function ComplexHeatmap::Heatmap().

show_row_dend

logical, default TRUE, indicating whether to show dendrogram in the heatmap or not. Internally this argument is passed to the function ComplexHeatmap::Heatmap().

row_names_font_size

a numeric value, default 10, indicating size of row names in the heatmap.

show_column_names

logical, default TRUE, indicating whether to show column names in the heatmap or not. Internally this argument is passed to the function ComplexHeatmap::Heatmap().

cluster_columns

logical, default TRUE, indicating whether to cluster columns in the heatmap or not. Internally this argument is passed to the function ComplexHeatmap::Heatmap().

show_heatmap_legend

logical, default TRUE, indicating whether to show heatmap legend or not.

...

Other parameters to be passes to ComplexHeatmap::Heatmap().

sample_comparisons

a character vector denoting sample comparisons for which heatmap of log2 fold-change to be plotted.

Value

an output of the function ComplexHeatmap::Heatmap().

Details

The function get_gene_expression_heatmap is to create a heatmap either for normalised gene expression or z-score values while the function get_fold_change_heatmap is to create a heatmap for log2 fold-changes values.

repair_genes : Internally gene names are stored as a "gene_id:gene_symbol" format. For example, "ENSG00000187634:SAMD11". When repair_genes is set to TRUE the string corresponding to gene_id followed by ":" will be removed. This is useful when gene names to be revealed in the heatmap.

convert_zscore : When set to TRUE values for each gene is converted into z-score. z-score is calculated by baser r function base::scale() with all default parameters.

Examples