Run DEseq2 for retained introns.

This function performs DEseq2 analysis for retained introns which are identified by IRFinder-S program. It mainly requires two inputs - 1) output file(s) from IRFinder-S program, and 2) sample group information to inform DEseq2 which samples to be grouped for differential expression analysis. Other arguments are to filter retained and differently expressed introns.

Usage

run_deseq_analysis_ir(
  x,
  min_intron_cov = 0.7,
  min_intron_depth = 10,
  minimum_splice_exact = 5,
  minimum_replicates_with_ri = 2,
  sample_info,
  group_numerator = "treatment",
  group_denominator = "control",
  cutoff_lfc = 1,
  cutoff_pval = 0.05,
  cutoff_padj = 0.01,
  regul_based_upon = 1
)

Arguments

x

an object of the class IRFinder-S. Usually this can be generated through the function read_irfinderS_output().

min_intron_cov

a value between 0 and 1, default 0.7, denoting minimum threshold for the intron coverage. Introns with coverage less than the value will not be considered for DE analysis.

min_intron_depth

an integer, default 10, denoting threshold for minimum number of aligned reads within an intron. Same value will be passed to argument min_counts in run_deseq_analysis() for differential analysis.

minimum_splice_exact

an integer, default 5, denoting threshold for minimum number of spliced reads within an intron.

minimum_replicates_with_ri

an integer, default 2, denoting threshold for number of replicates have an intron identified as "retained intron". Same value will be passed to argument min_replicates in run_deseq_analysis() for differential analysis.

sample_info

internally passed to run_deseq_analysis().

group_numerator

internally passed to run_deseq_analysis().

group_denominator

internally passed to run_deseq_analysis().

cutoff_lfc

internally passed to run_deseq_analysis().

cutoff_pval

internally passed to run_deseq_analysis().

cutoff_padj

internally passed to run_deseq_analysis().

regul_based_upon

internally passed to run_deseq_analysis().

Value

Examples

example_dir <- system.file("extdata" ,"ir_data" , package="parcutils")
example_files <- fs::dir_ls(example_dir, glob = "*-IR-nondir*.txt")
names(example_files) <- stringr::str_replace(example_files , pattern = ".*/", replacement = "") %>%
stringr::str_replace(pattern = "-IR-nondir.txt", replacement = "")
x <- read_irfinderS_output(files = example_files,  add_prefix_chr = FALSE)
ir_sample_info <- names(example_files) %>%
tibble::tibble(samples = . , groups = rep(c("control", "treatment"), each = 2) )
run_deseq_analysis_ir(x = x, sample_info = ir_sample_info,
min_intron_depth = 2, regul_based_upon =1, cutoff_pval = 0.01)
#> ℹ Running DESeq2 ...
#> converting counts to integer mode
#> estimating size factors
#> estimating dispersions
#> gene-wise dispersion estimates
#> mean-dispersion relationship
#> final dispersion estimates
#> fitting model and testing
#> ✔ Done.
#> ℹ Summarizing DEG ...
#> ✔ Done.
#> 
#> Total number of RI used for DE analysis are 31864. 
#> Total number of comparisons are 1. 
#> 
#> Number of DE RI in each comparison:
#> 
#> treatment_VS_control 
#> • number of up RI   : 251.
#> • number of down RI : 304.
#> ──────────────────────────────