Get a data matrix (PSI, logit, z-score) for a given ASE

Get a data matrix (PSI, logit, z-score) for a given ASE

Usage

get_ASE_data_matrix(
  se,
  event_names,
  samples,
  summarise_groups = TRUE,
  summarise_groups_by = "mean",
  method = "PSI",
  column_condition = "condition"
)

Arguments

se

an object of class NxtSE.

event_names

a character vector denoting valid ASE names to plot in the heatmap.

samples

a character vector denoting valid sample names to plot in the heatmap.

summarise_groups

logical, denoting whether to summaries groups (replicates mostly), default TRUE.

summarise_groups_by

a character string denoting a method to summaries groups. Can be one of the mean or median.

method

a character string denoting a method for data value. Can be one of the "PSI", "logit" or "Z-Score".

column_condition

a character string denoting a condition column from se to select samples. Same column will also be used to group and summaries samples.

Value

a data matrix

Examples

se <- SpliceWiz::SpliceWiz_example_NxtSE(novelSplicing = TRUE)
SpliceWiz::colData(se)$treatment <- rep(c("A", "B"), each = 3)
SpliceWiz::colData(se)$replicate <- rep(c("P","Q","R"), 2)
res <- run_ase_diff_analysis(x = se, test_factor = "treatment", test_nom = "A" ,test_denom = "B",  IRmode ="annotated",  cutoff_lfc = 0.6, cutoff_padj = 1, regul_based_upon = 2)
#> Mar 29 12:40:36 Performing edgeR contrast for included / excluded counts separately
#> Mar 29 12:40:37 Performing edgeR contrast for included / excluded counts together
event_names = get_ASEsets_by_regulation(x = res, sample_comparisons = "A_VS_B", regul = "all") %>% unlist()
get_ASE_data_matrix(se, event_names , samples = c("A", "B"), column_condition ="treatment")
#> # A tibble: 99 × 3
#>    event_name                                       A      B
#>    <chr>                                        <dbl>  <dbl>
#>  1 A3SS:NSUN5-201-exon3;NSUN5-203-exon3        1      0.996 
#>  2 A3SS:NSUN5-201-exon7;NSUN5-206-exon6        0.879  0.908 
#>  3 A3SS:NSUN5-201-exon8;NSUN5-novelTr003-exon2 0.953  0.958 
#>  4 A3SS:NSUN5-204-exon10;NSUN5-202-exon10      0.955  0.941 
#>  5 A3SS:SRSF1-201-exon4;SRSF1-novelTr001-exon2 0.766  0.877 
#>  6 A3SS:SRSF1-204-exon4;SRSF1-203-exon2        0.963  0.916 
#>  7 A3SS:SRSF2-202-exon2;SRSF2-201-exon2        0.988  0.989 
#>  8 A3SS:SRSF2-206-exon2;SRSF2-202-exon3        0.165  0.159 
#>  9 A3SS:SRSF2-206-exon2;SRSF2-novelTr003-exon2 0.988  0.981 
#> 10 A3SS:SRSF2-novelTr002-exon2;SRSF2-209-exon2 0.0530 0.0596
#> # … with 89 more rows
get_ASE_data_matrix(se, event_names , samples = c("P", "Q","R"), column_condition ="replicate")
#> # A tibble: 99 × 4
#>    event_name                                       P       Q      R
#>    <chr>                                        <dbl>   <dbl>  <dbl>
#>  1 A3SS:NSUN5-201-exon3;NSUN5-203-exon3        0.997  1       0.997 
#>  2 A3SS:NSUN5-201-exon7;NSUN5-206-exon6        0.836  0.911   0.933 
#>  3 A3SS:NSUN5-201-exon8;NSUN5-novelTr003-exon2 0.929  0.979   0.959 
#>  4 A3SS:NSUN5-204-exon10;NSUN5-202-exon10      0.954  0.957   0.933 
#>  5 A3SS:SRSF1-201-exon4;SRSF1-novelTr001-exon2 0.807  0.738   0.920 
#>  6 A3SS:SRSF1-204-exon4;SRSF1-203-exon2        0.970  0.984   0.865 
#>  7 A3SS:SRSF2-202-exon2;SRSF2-201-exon2        0.984  0.988   0.993 
#>  8 A3SS:SRSF2-206-exon2;SRSF2-202-exon3        0.158  0.219   0.110 
#>  9 A3SS:SRSF2-206-exon2;SRSF2-novelTr003-exon2 0.987  0.996   0.971 
#> 10 A3SS:SRSF2-novelTr002-exon2;SRSF2-209-exon2 0.0954 0.00864 0.0649
#> # … with 89 more rows