Run differential expression (DE) analysis for several comparisosns at a time. Internally it uses DESeq2::DESeq() to perform DE analysis.

DESeq2 is a popular method to perform DE analysis for RNA-seq data. Pre and post DESeq run, however, involves several data wrangling steps. For example, prior to run DESeq genes may need to be filtered out based on number of reads mapped to genes across replicates. Similarly, post DESeq run user needs to set cutoffs (log2fc, pvalue and padj) to define up and down regulated genes. For large RNA-seq experiments involving several DE comparison such as sub-setting and different cutoffs creates messy and less readable code. This function helps to make things bit tidy and increase the code readability. Besides that, output of this function can subsequently used for several other functions of this package.

Usage

run_deseq_analysis(
  counts,
  column_geneid,
  column_samples,
  sample_info,
  group_numerator,
  group_denominator,
  delim = "\t",
  comment_char = "#",
  min_counts = 10,
  min_replicates = 1,
  cutoff_lfc = 1,
  cutoff_pval = 0.05,
  cutoff_padj = 0.01,
  regul_based_upon = 1,
  ...
)

Arguments

counts

a character string of the path to a count file or an object of dataframe having raw counts for each gene. See details below to know more about format of the count file and count dataframe.

column_geneid

a character string denoting a column of geneid in counts

column_samples

a character vector denoting names of sample columns from counts

sample_info

a character string denoting a name of sample information file or a dataframe. A file or a dataframe both must have at least two columns without column names. First column denotes to samples names and second column denotes group name for each sample in first column. For e.g.

sample 1	WT

sample 2	WT

sample 3	KO

sample 4	KO

group_numerator

a character vector denoting sample groups to use in numerator to calculate fold change.

group_denominator

a character vector denoting sample groups to use in denominator to calculate fold change.

delim

a character denoting deliminator for count file. Only valid if count is a file path.

comment_char

a character denoting comments line in count file. Only valid if count is a file path.

min_counts

a numeric value, default 10, denoting minimum counts for a gene to be used o consider a gene for differential expression analysis.

min_replicates

a numeric value, default 1, denoting minimum samples within a group must have minimum_counts. Value provided must not be higher than number of replicates in each group. For example for given values min_replicates = 2 and minimum_counts = 10 the genes which have minimum counts 10 in at least 2 sample groups will be used for DEG.

cutoff_lfc

a numeric value which is internally passed to categorize_diff_genes

cutoff_pval

a numeric value which is internally passed to categorize_diff_genes

cutoff_padj

a numeric value which is internally passed to categorize_diff_genes

regul_based_upon

either of 1, 2 or 3 which is internally passed to categorize_diff_genes

...

for future use

Value

a data frame of DESeq results, DEG, and DEG summary.

Details

// TO DO : Explain formats of count file and count dataframe.

Examples

if (FALSE) {

count_file <- system.file("extdata","toy_counts.txt" , package = "parcutils")
count_data <- readr::read_delim(count_file, delim = "\t")

sample_info <- count_data %>% colnames() %>% .[-1]  %>%
 tibble::tibble(samples = . , groups = rep(c("control" ,"treatment1" , "treatment2"), each = 3) )

res <- run_deseq_analysis(counts = count_data ,
                         sample_info = sample_info,
                         column_geneid = "gene_id" ,
                         group_numerator = c("treatment1", "treatment2") ,
                         group_denominator = c("control"),
                         column_samples = c("control_rep1", "treat1_rep1", "treat2_rep1", "control_rep2", "treat1_rep2", "treat2_rep2", "control_rep3", "treat1_rep3", "treat2_rep3"))

res

## all comparisons

print(res$comp)

## DESEq result object(s)

print(res$dsr)

## DESEq result data frame

print(res$dsr_tibble)

## DESEq result data frame  DEG assigned, look at the columns 'signif' and 'regul'

print(res$dsr_tibble_deg)

## DEG summary

print(res$deg_summmary)

}